+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to the "Sharpen" extension will be documented in this file.
 
 The format of the file is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased]
+### Added
+- "Use out variables in method invocations" suggestion.
+
 ## [0.2.0] - 2017-12-14
 ### Added
 - "Use default expression in optional constructor parameters" suggestion.